bons10 - FIxed the seeding problem
